pyproject.toml
[project]
name = "scada-agent"
version = "0.1.0"
description = "Integração SCADA-LTS com Agente Inteligente (Claude/Gemini)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.28.0",
    "pandas>=1.5.0",
    "openpyxl>=3.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.7.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.39.0",
    "websockets>=15.0.1",
    "httpx>=0.28.1",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
]
